requirements.txt
# Backend API Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Data validation
pydantic==2.6.1

# Password hashing
passlib[bcrypt]==1.7.4

# JWT Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9

# CORS support (included with FastAPI but listed for clarity)
# fastapi already includes python-multipart and email-validator
python-dotenv==1.0.0